Organized GUI Github Layout
