+++ Images/_20241024_163141/frames_0008.bin
