+++ Images/_20241024_164417/frames_0008.bin
